% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_sum.R
\name{permutation_sum}
\alias{permutation_sum}
\title{A function that help you make get the overlap of SE and permuted CNA excluding chromosome X and Y}
\usage{
permutation_sum(data1, data2, cnadata)
}
\arguments{
\item{data1}{Super Enhancer Sample One information about chromosome name, start and end position}

\item{data2}{Super Enhancer Sample Two information about chromosome name, start and end position}

\item{cnadata}{A list of permutation CNA results}
}
\description{
This function allows you to make get the overlap of SE and permuted CNA
}
\examples{
permutation_sum()
}
\keyword{cnVpermtest}
